Add Class FormatterInfo and OtherInfoConstants
Se agregaron dos clases extras para formatear la información y que la vista se libere de dicha tarea y solo tenga que mostrarla.
