+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/presentation/OtherInfoConstants.kt
@@ -1,0 +1,7 @@
+package ayds.lisboa.songinfo.moredetails.fulllogic.presentation
+
+const val HTML_WIDTH = "<html><div width=400>"
+const val HTML_FONT = "<font face=\"arial\">"
+const val HTML_END = "</font></div></html>"
+const val NO_RESULTS = "No results"
+const val PREFIX_LOCALLY_STORED = "[*]"
